[Mac] do not create a dummy source object when property values are requested (#1587)
* do not create a dummy source object

* do not create an obs_source here because frontend (SLD) is currently invoking this
function before obs_init_graphics() has been invoked. Typically that is fine
however, if we want to use a plugin like macos_avcapture_fast, then
we need for video to be init BEFORE creating the source (because it will in turn, attempt to
locate a shader effect). Also, I do believe this
code to be more efficient and more closely replicates what OBS-Frontend does
when it lists properties for a source

* Add warning messages if we fail to list properties for a source

* set frame_rate for macos_avcapture_fast

* replace strcmp with std::string_view to search for substring