release the obj for all source types
* not sure if intentional? if a source type wasnt type Transition or scene,
the source would be removed immediately ignoring the source ref count. Changing
this code will make the code properly respect the ref count. Only red flag is
I cannot be 100% sure if the author did this on purpose or was it a typo? I'm
assuming its a typo because at first glance, release() and remove() both seem
like they might do the samething under the hood (but they're actually a bit different)