check for bin folder on Apple
* On Windows, the cmake file doesnt install obs64
into bin folder like Apple so updating this logic
to be a little more explicit. Resolves issue I
encountered on Windows when I accidently put a bin folder
in the desktop/node_modules/obs-studio-node parent
folder which caused 'yarn run test' to FAIL to spin up
obs64 ipc server.

Below is error feedback added if obs64 exe is missing:
* Adding new exitcode that represents the obs64 exe
could not be found. Updating JS code to properly
check the exitCode to avoid a silent failure case
* Output to error stream when connection fails.
Previously, the code attempted to return a JS
exception but it never got to run due to the exit(1).
This way we maintain legacy behavior and exit application
with a helpful message