add error feedback if obs64 exe is missing
* controller.cpp: Adding new exitcode that represents
the obs64 exe could not be found. Updating JS code
to properly check the exitCode to avoid a silent
failure case
* utility.hpp: Output to error stream when connection fails.
Previously, the code attempted to return a JS
exception but it never got to run due to the exit(1).
This way we maintain legacy behavior and exit
application with a helpful message
* module.js: verify bin folder contains obs64